refactor: 상품 상세 페이지 Header 컴포넌트 생성 (#171)
- ProductDetailMain에 있던 Header를 분리
