+++ fe/src/components/ProductDetail/ProductDetailHeader/style.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const Header = styled.header`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: fixed;
+
+  width: 100%;
+  height: 44px;
+  padding: 16px;
+`;
+
+export { Header };
